return inserted value when calling put in lru cache
This fixes the fact that item inserted in the cache were returned by
value, thereby creating a spurious copy of something which was in the
cache.

The put method on the lru cache now return a reference to the inserted
value. This allow us to return the ref without having to perform a key
lookup.